Average / Students

+++ Trabajo_Promedio/average.py
@@ -1,8 +1,8 @@
 #Se Solicita el ingreso de las cuatro notas.
-c1 = int(input("1: "))
-c2 = int(input("2: "))
-c3 = int(input("3: "))
-c4 = int(input("4: "))
+c1 = float(input("1: "))
+c2 = float(input("2: "))
+c3 = float(input("3: "))
+c4 = float(input("4: "))
 
 #Se suman las variables anteriores y el resultado
 #Se almacena en la variable s.
@@ -14,4 +14,3 @@ p = s / 4
 
 #Se imprime la salida.
 print(f"Promedio: {p}")
-
